src/backend/requirements.txt
Flask
mysqlclient
flask-mysqldb
python-dotenv
requests
Flask-Cors
Flask-Session2